requirements.txt
streamlit
langchain_openai
requests
langchain
langchain_community
datasets
openai
numpy
transformers
torch
sentence_transformers
langchain_huggingface
ragas
weaviate-client
streamlit_feedback
pinecone-client
langchain_pinecone
langchain_weaviate
langsmith
langgraph
pandas
scipy
pillow
torchvision
sentence-transformers
unidecode
pytesseract
langchain_mistralai
pymupdf
langchain-huggingface
llmlingua
accelerate
pyarrow
lancedb
pillow_heif
llama-index-vector-stores-lancedb
llama-index
ftfy
tqdm
llama-index-multi-modal-llms-openai
llama-index-embeddings-huggingface
llama-index-readers-file
einops
unstructured
unstructured_inference
unstructured.pytesseract
pdfminer
llama-index-embeddings-clip
scikit-image
scikit-learn
matplotlib
Spire.Pdf